feat: user middleware
- Add user middleware to fetch user data from db.
- Update other parts of the code accordingly.
